Refactor db.search to use memory
Tried a number of different search things: PouchDB-find, Pouch Map
function, raw IndexedDB and nothing seemed to be working quickly.  Now
I just create a hash map for both generic name and NDC so we get O(1)
lookups.  Memory is an issue.  Generic name hash map for 30K drugs
seems to be around 65Mb and NDC hash map seems to be 30Mb.  Since we
are closer to 90K drugs in actual database having all them in memory
might not be feasible.